refactor(challenges): move pure selectPayableUsers to db-free module so unit tests load
challenge-rewards.ts imports the prisma client; importing it from a unit test
triggers env.LOGGING.filter under vitest. Extract the pure helper to
challenge-rewards.utils.ts (re-exported for API parity) and point the test there,
matching the project convention that unit tests import only db-free modules.
